requirements.txt
flask
flask-cors
whisper
openai
langchain
pinecone-client
